refactor: split canvas editor into reusable components
