+++ src/components/IconButton.jsx
@@ -7,6 +7,7 @@ const IconButton = ({ onClick, title, children, className = "" }) => {
       onClick={onClick}
       title={title}
       className={`p-2 rounded bg-white shadow hover:bg-blue-100 ${className}`.trim()}
+      type="button"
     >
       {children}
     </button>
